Made path following more accuate; merged Christoper's and my code together.

+++ hkust-robocon-mb/stm32f103VF_template/src/user/pid.c
@@ -5,14 +5,14 @@
 #include "pid.h"
 #include "approx_math.h"
 
-void updatePid(PID *pid, int32_t error) {
+void updatePid(PID *pid, float error) {
 	pid->errorRate = error - pid->error;
 	if ((pid->totalError + error) * pid->Ki / pid->scale < pid->max)
 		pid->totalError += error;
 	pid->error = error;
 	
 	if (Abs(pid->error) > pid->threshold)
-		pid->output = ((pid->error * pid->Kp) + (pid->totalError * pid->Ki) + (pid->errorRate * pid->Kd)) / pid->scale;
+		pid->output = ((pid->error * pid->Kp) + (pid->totalError * pid->Ki) + (pid->errorRate * pid->Kd)) / (float) pid->scale;
 	
 	if (pid->output > pid->max) pid->output = pid->max;
 	if (pid->output < pid->min) pid->output = pid->min;
